Finish initial setup
