+++ package.json
@@ -2,13 +2,23 @@
   "name": "@shgysk8zer0/postcss-import",
   "version": "1.0.0",
   "description": "Plug-in for importing local & remote stylesheets with PostCSS ",
+  "keywords": [
+    "postcss",
+    "postcss-plugin",
+    "postcss-import",
+    "postcss-import-url",
+    "@import"
+  ],
   "type": "module",
-  "main": "index.cjs",
-  "module": "index.js",
+  "main": "plugin.cjs",
+  "module": "plugin.js",
   "exports": {
     "./": {
-      "import": "./index.js",
-      "require": "./index.cjs"
+      "import": "./plugin.js",
+      "require": "./plugin.cjs"
+    },
+    "./*": {
+      "import": "./*.js"
     }
   },
   "engines": {
@@ -40,6 +50,16 @@
   "bugs": {
     "url": "https://github.com/shgysk8zer0/postcss-import/issues"
   },
+  "funding": [
+    {
+      "type": "librepay",
+      "url": "https://liberapay.com/shgysk8zer0"
+    },
+    {
+      "type": "github",
+      "url": "https://github.com/sponsors/shgysk8zer0"
+    }
+  ],
   "homepage": "https://github.com/shgysk8zer0/postcss-import#readme",
   "devDependencies": {
     "@shgysk8zer0/js-utils": "^1.0.0",
@@ -49,5 +69,8 @@
   },
   "dependencies": {
     "@shgysk8zer0/npm-utils": "^1.0.5"
+  },
+  "peerDependencies": {
+    "postcss": "*"
   }
 }
