+++ postcss.config.js
@@ -1,9 +1,8 @@
 /* eslint-env node */
-// import pcImport from 'postcss-import';
-// import pcURL from 'postcss-url';
-// import pcImportURL from 'postcss-import-url';
-import { postCSSImport } from './index.js';
+import postCSSImport from './plugin.js';
+
 export default {
-	map: {inline: false},
-	plugins: [postCSSImport({ resolveUrls: false, recursive: false })]
+	// map: { inline: false },
+	// plugins: [postCSSImport({ resolveUrls: false, recursive: false })]
+	plugins: [postCSSImport]
 };
