+++ rollup.config.js
@@ -1,10 +1,8 @@
 import { getConfig } from '@shgysk8zer0/js-utils/rollup';
 
-export default getConfig('./index.js', {
+export default getConfig('./plugin.js', {
 	format: 'cjs',
 	minify: false,
 	sourcemap: false,
 	external: ['postcss', 'node:fs/promises'],
-	plugins: [],
-	globals: {},
 });
